+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+- Added a new named constructor to initialize `Decimal` to zero.
+- Improved `toString()` method to remove trailing zeros.
+
 ## 0.0.1
 
 - Initial version.